Added run as root requirement on scripts.

+++ servidor4/netplan/network.sh
@@ -11,6 +11,11 @@ if [ "$SCRIPT_PATH" != "$CURRENT_PATH" ]; then
 	exit 1
 fi
 
+if [ "$EUID" -ne 0 ]
+  then echo ERROR! Please run as root. 1>&2
+  exit
+fi
+
 rm /etc/netplan/* 2> /dev/null
 chmod 600 network.yaml
 cp network.yaml /etc/netplan/
